News-ticker/requirements.txt
google-adk[database]==0.3.0
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
deprecated==1.2.14
flask==3.0.1
flask-cors==4.0.0
requests==2.31.0
pydantic==2.6.1
pydantic-core==2.16.2
